сохранение смещения в mmap

+++ kernel/proc/syscalls.h
@@ -82,7 +82,7 @@ int sys_thread_sleep(time_t tv_sec, long int tv_nsec);
 
 int sys_pause();
 
-void* sys_memory_map(void* address, uint64_t length, int protection, int flags, int fd, int offset);
+void* sys_memory_map(void* address, uint64_t length, int protection, int flags, int fd, off_t offset);
 
 int sys_memory_protect(void* address, uint64_t length, int protection);
