+++ sdk/libc/src/sys/mman.c
@@ -3,7 +3,7 @@
 #include "stdint.h"
 #include "syscalls.h"
 
-void* mmap(void* addr, size_t length, int protection, int flags, int fd, int offset)
+void* mmap(void* addr, size_t length, int protection, int flags, int fd, off_t offset)
 {
     void* result = syscall_map_memory(addr, length, protection, flags, fd, offset);
     int64_t rc = (int64_t) result;
